# drustvena_mreza/Web App/api/api.yaml
swagger: '2.0'
info:
  title: Bubbles API
  description: Mrzim svoj život.
  version: 1.0.0
produces:
  - application/json
paths:
  /isUsernameAvailable:
    get:
      tags:
        - Android
      summary: Provjerava ima li registrirani korisnik s tim imenom
      parameters:
        - name: username
          type: string
          in: query
          required: true
      responses:
        '200':
          description: true ili false
          schema:
            $ref: '#/definitions/Available'
  /isEmailAvailable:
    get:
      tags:
        - Android
      summary: Provjerava ima li registrirani korisnik s tim e-mailom
      parameters:
        - name: email
          type: string
          in: query
          required: true
      responses:
        '200':
          description: true ili false
          schema:
            $ref: '#/definitions/Available'
  /sign-up:
    post:
      tags:
        - General
      summary: Registracija
      parameters:
        - name: username
          in: formData
          required: true
          type: string
        - name: email
          in: formData
          required: true
          type: string
        - name: password
          in: formData
          required: true
          type: string
        - name: firstName
          in: formData
          required: false
          type: string
        - name: lastName
          in: formData
          required: false
          type: string
        - name: middleName
          in: formData
          required: false
          type: string
        - name: address
          in: formData
          required: false
          type: string
        - name: city
          in: formData
          required: false
          type: string
        - name: country
          in: formData
          required: false
          type: string
      responses:
        '200':
          description: Kod uspješne registracije.
        default:
          description: Vraća kolekciju pogreški vezane uz poejdino polje
          schema:
            $ref: '#/definitions/ValidationError'
  /sign-in:
    post:
      tags:
        - General
      summary: Prijava
      parameters:
        - name: username
          in: formData
          required: true
          type: string
        - name: password
          in: formData
          required: true
          type: string
      responses:
        '200':
          description: Uspješna prijava
        default:
          description: Poruka vezana uz neuspješnu prijavu.
          schema:
            $ref: '#/definitions/Error'
...
  /bubble/create:
    post:
      tags:
        - Bubble
      summary: Stvara novi bubble za aktualnog vlasnika session-a.
      parameters:
        - name: title
          in: formData
          required: true
          type: string
        - name: description
          in: formData
          required: false
          type: string
      responses:
        '200':
          description: Uspješno spremanje u bazu. Vraća netom stvoreni bubble.
          schema:
            $ref: '#/definitions/Bubble'
        '400':
          description: Nevaljani formData.
          schema:
            $ref: '#/definitions/ValidationError'
  '/bubble/edit/{id}':
    post:
      tags:
        - Bubble
      summary: Omogućava promjenu naslova i opisa postojećeg bubble-a.
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - name: title
          in: formData
          required: true
          type: string
        - name: description
          in: formData
          required: false
          type: string
      responses:
        '200':
          description: Uspješna promjena podataka. Vraća netom izmijenjeni bubble.
          schema:
            $ref: '#/definitions/Bubble'
        '400':
          description: Nevaljani formData.
          schema:
            $ref: '#/definitions/ValidationError'
        '403':
          description: Pokušaj promjene tuđeg bubble-a.
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Pokušaj promjene nepostojećeg bubble-a.
          schema:
            $ref: '#/definitions/Error'
  '/bubble/{id}':
    post:
      tags:
        - Bubble
      summary: Dohvat contenta bubble-a sa id-em id.
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: Dohvaćeni contenti.
          schema:
            $ref: '#/definitions/Content'
...
definitions:
...
  Bubble:
    type: object
    properties:
      user_id:
        type: integer
      id:
        type: integer
      bubble_type_id:
        type: integer
      created_at:
        type: string
      updated_at:
        type: string
      title:
        type: string
      description:
        type: string
  Content:
    type: object
    properties:
      id:
        type: integer
      bubble_id:
        type: integer
      content_type_id:
        type: integer
      created_at:
        type: string
      updated_at:
        type: string
      title:
        type: string
      content:
        type: string
...
  Available:
    type: object
    properties:
      available:
        type: boolean
  Error:
    type: object
    properties:
      error:
        type: string
  ValidationError:
    type: object
    properties:
      errors:
        type: array
        items:
          type: string